+++ src/Lotto.js
@@ -1,4 +1,4 @@
-import { MESSAGES, LOTTO_NUMBER, PRIZE_MONEY } from "./constants/index.js";
+import { MESSAGES, LOTTO_NUMBER, PRIZE_MONEY } from "./constants/lotto.js";
 import { MissionUtils } from "@woowacourse/mission-utils";
 
 class Lotto {
