+++ src/constants/lotto.js
